+++ libs/notifications/src/index.ts
@@ -12,7 +12,7 @@ assert(queueUser, 'QUEUE_USER is required');
 const queuePassword = process.env.QUEUE_PASSWORD;
 assert(queuePassword, 'QUEUE_PASSWORD is required');
 
-export interface Notification {
+export interface PushNotification {
   id: string;
   account: string;
   title: string;
@@ -22,11 +22,13 @@ export interface Notification {
 
 export function parseNotifications(
   notificationsString: string
-): Notification[] {
+): PushNotification[] {
   return JSON.parse(notificationsString);
 }
 
-export function stringifyNotifications(notifications: Notification[]): string {
+export function stringifyNotifications(
+  notifications: PushNotification[]
+): string {
   return JSON.stringify(notifications);
 }
 
@@ -67,7 +69,7 @@ export async function connectToChannel(
 export interface SendToQueueParams {
   channel: Channel;
   queue: string;
-  notifications: Notification[];
+  notifications: PushNotification[];
 }
 
 export function sendNotificationsToQueue(params: SendToQueueParams) {
